packages/backend/pyproject.toml
[build-system]
requires = ["setuptools==75.8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vaultguard-backend"
version = "1.0.0"
description = "VaultGuard backend service"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.8",
  "uvicorn[standard]==0.34.0",
  "sqlalchemy==2.0.38",
  "aiomysql==0.2.0",
  "argon2-cffi==23.1.0",
  "alembic==1.14.1",
  "pydantic-settings==2.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.3",
  "httpx==0.28.1",
  "aiosqlite==0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
